[factory] clean and separate into files
