+++ frostsnap_factory/Cargo.toml
@@ -19,3 +19,4 @@ frostsnap_coordinator.workspace = true
 tracing.workspace = true
 tracing-subscriber.workspace = true
 bincode.workspace = true
+rusqlite = { version = "0.31.0", features = ["bundled"] }
